Strip out gamepad wasm code
